Add ":index flow <dir>"
Summary:
It's really useful for experimenting to be able to just index a set of
source files in the shell and then run queries against the data. Doing
this manually is a bit of a pain, so I've automated it, for flow only
so far.

Caveats:
* Only for flow
* Doesn't derive any predicates. That would be relatively easy to add,
but keeping the set of predicates in sync will be tricky.

Could we do this for other languages?
* Hack and Java: these both produce JSON, so should be as easy as Flow
in theory.
* Python: the shell would need to spin up a local server for the
indexer to talk to
* C++/Objective C: not hard, we have all the pieces in the regression
framework

Reviewed By: zsol

Differential Revision: D26979918

fbshipit-source-id: fa27c33bd82b55c9032e9b55b554a6f1d9cdd9b2