Enable returning result count but not the results themselves
Summary:
Add flag `count_results` to `UserQueryOptions` allowing results to be omitted in the response and only their count be returned.

This saves the server from encoding the query results and from sending them through the wire, which speeds up the case where one does not need the query results. This is exactly the case when using `glean derive`.

The flag is enabled in `glean derive` to save encoding values that would be ignored by the client.

Reviewed By: simonmar

Differential Revision: D26982172

fbshipit-source-id: 434b982d1e63c1b7fed07784a703d0001cdd33d6