Handle references from imports properly
Summary:
We were relying on their being reflexive references for the symbols in
an import statement. According to vrama628 on T86746757 this
is unintentional, and in fact there are no reflexive references for
type names.

In this diff we generate DirectXRefs for import declarations
explicitly, without relying on the reflexive references.

Reviewed By: nhawkes

Differential Revision: D27078836

fbshipit-source-id: dbb875a588c1fcc595bed1337a23c52494f71f55