Add docmarkup.angle schema
Summary:
# What is the goal?

Support use cases of annotations in comments (e.g. fuzzying and UI explorer) with Glean facts about parsed documentation comments.

Full project plan is https://fb.quip.com/koHAAVaPPPdB

# What was not the way?

Codex (in CodexDocblockParser) pases comments (into CodexDocblock) included the the special map of `keys` to values.

These were used in some UI explorer. The Lionhead fuzzying needs similar annotations and was going to invent their own indexer.

# What is the new way?

Make a Glean indexing pass to load and parse documentation comments.

This diff adds the docmarkup.angle schema to hold and lookup the parsed facts.

# Size estimate?

A rough estimate of sizing:

For our first target of fbsource, the size of the cxx1.DeclarationComment is:
```
cxx1.DeclarationComment.1
  count: 892244
  size:  43170891 (41.17 MB) 0.0348%
```
So the new facts will be relatively small.

For www the current sources are:
```
hack.DeclarationComment.4
  count: 18763183
  size:  1006505117 (959.88 MB) 2.2765

hack.UserAttribute.4
  count: 823360
  size:  155555005 (148.35 MB) 0.3518%
```
Estimate how many parameters we will find in the hack.DeclarationComment.4 :
```
$ ~/bins/biggrep_client biggrep.master www re2 --count -e '^ *\* *@ [a-zA-Z]'
8768363 hits for this query
```
So the UserAttribute impact is small, but the DeclarationComment parsing will be a few percent.
So we expect 9 million -tags in 18 million comments.

Reviewed By: simonmar

Differential Revision: D26512769

fbshipit-source-id: 50a2d02efc2eb6460b04b92ee83f79097e2fdba9