Make the shell and server independent of the schema
Summary:
Really the server and the shell should both be independent of the schema, but for legacy reasons they currently aren't. Let's fix these dependencies for the open source build at least.

Server: The search Thrift API will go away in due course.
Shell: The magic xrefs query for investigating C++ xrefs is a horrible hack that should probably be a derived predicate or something.

Reviewed By: zsol

Differential Revision: D27156748

fbshipit-source-id: f4ca062ae9825600f5ad18796ed27b1d69abd562