Use the "all" schema for resolving unversioned queries
Summary:
We want to make it easier to do schema migrations, and right now one
problem is that as soon as you land a new schema, all unversioned
queries will target the new schema. To be able to do things more
incrementally, we want to be able to control when that happens.

We already have the "all" schema to control which schema versions are
used for codegen, the change in this diff is that we also use it to
control which schema versions are used to resolve unversioned queries.

Reviewed By: ChrisKuklewicz

Differential Revision: D27267778

fbshipit-source-id: ae62da75ed12674c14236a7a909f26a04bfda505