Better "reference" codegen
Summary:
# What is the goal?

Codegen from angle should compile.

# What was not the way?

There were problems due to cpp.ref thrift annotations, making cython fail and later lionhead.lionizer fail.

# What is the new way?

`needsRefType`  can fail to resolve the underlying representation type, because it is not in the environment.  In this case it must be from some imported module

# Details before I forget:

Let me explain the bug:

There was investigation done in D27044773, with Debug.Trace and debug.angle example schemas.

These were triggered by fields that were typedefs to `string`, where the typdef was imported in angle.

The root cause was that `needsRefType` was returning True when in fact it does not know the real type.  Thus the unknown cases were turned into references. This created failing code generated in C++ when `string` was the underlying type.

Should fbthrift be smarter? Perhaps.  Should we annotate `string` this way? No.  Should the environment have been more complete and so `needsRefType` would see the `string`? Probably.

Reviewed By: simonmar

Differential Revision: D27297344

fbshipit-source-id: 1d73c9b45cc0413b627263f4f0bfe57f12934ae1