Better nested-query reordering optimisation
Summary:
We have encountered several examples where the Angle compiler was
doing a poor job of ordering nested matches. This diff should fix a
large class of them.

The main change is to track which statements will become lookups; that
is, a statement of the form `X = pred P`, where X is known. We know
that each of the variables in P will be bound (cheaply) by this
statement, so we can treat statements with those variables on the
left-hand-side as lookups, and so on.

I've also made a slight tweak to the property "irrefutable", weakening
it to mean "full scan of all facts", which is a better approximation
to the property we want (an expensive search).

Reviewed By: nhawkes

Differential Revision: D27591560

fbshipit-source-id: 8dd91f9ac86090c088f061d445b33bf22605a7d8