Predicate to filter class dependencies from static method XRefs
Summary:
# What
Introduce a new Glean predicate in the `omegaanalyser` schema to support a new filter that removes class xrefs that exist only because of static method calls. The filter is introduced in D27533333.

**More context**
OA runs recursive searches where each "round" it does multiple queries equivalent to: **"Given a source node, get me all the nodes that it depends on"**. The source node can be a class, a function or a method. The result of this query will be a list of nodes (of mixed types listed above) that are dependencies of the source node.

These dependencies are obtained by using cross-references. If method `A::foo` calls method `B::bar`, then the method node `A::foo` depends on the method node `B::bar`. If `A::foo` instantiates the class `RandClass`, that means that the method node `A::foo` also depends on the **class node** `RandClass`.

**On the context of Omega Analyser / Experiments: X depends on Y if changing Y can affect X**.

## Problem
The way that the Hack indexer generates XRefs facts (which are stored in the Hack Glean DB and used to generate OA's dependency map) doesn't always map XRefs to dependencies **in OA's context** directly.

If a method `A::foo` calls a static method `B::staticBar`, the Hack indexer will generate 2 XRefs: one from `A::foo` to `B::staticBar` (the one we want, because it's a dependency) and **one from `A::foo` to `B`, which we don't want**.

We don't to map the 2nd XRef to a dependency in the DepMap because it means that when someone modifies **anything in the class `B`**, we'll consider that `A` could be affected. This is an over-approximation (?) that leads to **a lot of of false positives**.

**Note:** it's important to state that we always keep the 1st dependency (to the static method node), which is another reason why the 2nd one is not needed.

# How to solve this

For every **class node** that I get back in each search, I'll run a filter to make sure that it wasn't returned as a dependency only because the source node called its static methods. I'll do that by counting the number of XRefs/uses from the source node to (a) the class node and (b) any static method from the class node. If the values of (a) and (b) are the same, it means that the class should be **filtered out**.

# Example
```
class A {
  function randomMethod() {
    B::foo();
    ...
    B::bar();
    ...
    $instance = new C();
  }
}

class B {
  static function foo() {
    ...
  }

  static function bar() {
    ...
  }
}

class C {
  ...
}
```

In this case, we'll get 2 XRefs to **the class `B`**: one for the call to `B::foo` and one for the call to `B::bar`. We **don't want** these XRefs to **the class `B`**, because we already have XRefs to the **methods** `B::foo` and `B::bar`.

The XRef to class `C`, on the other hand, we **definitely want to keep**, because it maps to a dependency from class `A` to class `C`, since it's instantiating (and using) `C`.

# This Diff
Introduce a helper predicate that makes writing the filter in Haskell much easier. This predicate will take (a) a source node and (b) a class node (`targetClass`). It will return (c) all the uses of `targetClass` and (d) all the uses of static methods from `targetClass` in the source node.

Reviewed By: simonmar

Differential Revision: D27264394

fbshipit-source-id: e410e6cd7b96abde8e15e9656fd51d0dca974a07