Add support for query performance tests
Summary:
The basic idea is that we want a way to write a test for an Angle
query against some actual code that we've indexed, and record its
performance characteristics so that we can catch regressions.

For "performance characteristics" here we'll just use the number of
fact searches performed, i.e. the same as the output of `:profile
full` in the shell.

To write a performance test:

* Write a `foo.query` file as before
* Set `perf: true` in the file
* The performance results will go into `foo.perf`
* Run with `--replace` as before to generate or update the results

Reviewed By: ChrisKuklewicz

Differential Revision: D27857496

fbshipit-source-id: 19e63019c32649567d3a99a38381eb5a7ddb1baa