Yet more reordering
Summary:
I tried to build a reordering pass that would reorder everything, in
the hope that we could do a reasonable job. After a couple of days of
hacking, I ran into a case where the optimiser had to make a decision
that it couldn't make without knowledge of the data - I was betting
that this wouldn't happen with our code, and if it did we could use
the hints that the programmer supplied in the form of the initial
ordering of statements - but that turned out to be false.

So, what we have here is something much less ambitious. We'll do
reordering when
* there are nested predicate matches, as before
* we can hoist or sink simple O(1) statements (the new bit)

This helps in the case that zsol recently ran into (D27736532 (https://github.com/facebookincubator/Glean/commit/3719e509bc43f37d6cf518456d6e86fbb1d62fa0)) and
should also fix most cases of "cannot resolve".

In due course we might expand this reordering phase to be more clever,
but it's clear to me at this point that we want the human to have
plenty of control over query optimisation, so I'm keeping things
conservative for now.

Reviewed By: zsol

Differential Revision: D27823679

fbshipit-source-id: 0df512934b68e6648aa485d0d9076cd8dd78fabb