Implement derivePredicate and pollDerivation
Summary:
Move predicate derivation control to the server.
Whilst previously the client would perform multiple userQueries until a stored
predicate was completely derived, now this functionality is being moved to the
server. The first step is the implementation of the endpoints `derivePredicate`
and `pollDerivation`, which is the content of this diff.

The client will trigger the derivation process by calling `derivePredicate` and be able to
query for its progress with `pollDerivation`.

Reviewed By: simonmar

Differential Revision: D27867093

fbshipit-source-id: 567656239939ba09b581ce679f6687462a7e7fc8