Add derived predicate that maps XRef => File
Summary:
This diff adds a derived predicate that maps an `XRef` to the `File` that it belongs to. The derived predicate will significantly speed up queries to find the File owner of an XRef (helpful in detecting dead fields).

## Context
In order to detect dead fields in the type models, I've written a script that checks each getter[0] on a type model. If that getter is unused than we can remove the field from the type model schema.

Originally I assumed that dead fields would be completely unused and would have 0 glean XRefs that target them. However, it turns out that dead fields are not completely unreferenced in code; dead type model fields actually have a handful of references from other generated type model code. For example, Type Model's that have generated builders include a call to each field getter method. Additionally, the generated type model to flatbuffer  converter also calls each field getter for a type model.

This means that determining if a field is dead requires more than just counting `XRef`s to a method. We also need to lookup which file each `XRef` belongs to: *Iff a getter only has `XRef`s owned by files in a `generated-graphql` directory, then we can consider it dead*.

[0] in this case one getter is equivalent to one field in `type-schema.graphql`

Reviewed By: simonmar

Differential Revision: D27967665

fbshipit-source-id: ba3033b5ec7979c03c339f579f80934970bd2ef3