Add a Rust schema for Glean.
Summary:
This adds a schema for Rust support to the Glean indexer. It's essentially a
translation of the information that `rustc` emits with `-Z save-analysis`,
which also powers the RLS and the `rust-analyzer` tool. A conversion tool from
the `-Z save-analysis` format to this Thrift schema has been written, and I
intend to land it as a follow-up. The goal is to power basic jump-to-definition
for diffs and VS Code when macros aren't involved.

Reviewed By: simonmar

Differential Revision: D27898124

fbshipit-source-id: e2301bc0b3e7f03b61c67201d0542a3a5b637baa