Add flow.DeclarationUses and codemarkup.FlowEntityUses
Summary:
Wire up find-references for Flow in the codemarkup backend.

Undo the codemarkup.FileEntities perf hack while we're here.

Size impact on storing this:
```
www.flow> :stat
flow.DeclarationUses.3
  count: 39001891
  size:  1833419605 (1.71 GB) 100.0000%

Total size: 1.71 GB
```

But we dont' need to because perf is fine:

```
## find-references-by-location ...

[WWW/Flow] Found 148 refs to 50 defs in average 8.31ms
[Flow 76k] Found 202 refs to 50 defs in average 24.29ms
```

(with schema override on localhost, 8.3ms for find-refs by location is ok).

Reviewed By: simonmar

Differential Revision: D28059046

fbshipit-source-id: 8cc579e53b3a449abb5d75e4857521202b64a563