Fix a query optimisation bug, and add a test
Summary:
The reorder pass was looking at these two statements:

```
  F_3:testinfra.CoveredFile.3 = testinfra.CoveredFile.3 {_0,_};
```

```
  _10:testinfra.CoveredFileByTagAndAssembly.4 =
    testinfra.CoveredFileByTagAndAssembly.4<- ({F_3,T_4,A_5})
```

and decided that the first should come after the second, on the
grounds that `F_3` is not in a prefix position, which is true except
that the second statement here does not bind `F_3` (it only builds a
fact of the derived predicate) so we can't turn the first statement
into a lookup by placing it after the second statement.

The fix is to be more precise about which variables we consider to be
"bound in a non-prefix position".

Reviewed By: ChrisKuklewicz

Differential Revision: D28120730

fbshipit-source-id: e3e662e5b0b583df95bb0db938ad5b6448fc00f0