disable concurrent and complete predicate derivation check
Summary:
Hack deriver for aros.xros.flow, xplat.js.flow, www.flow and www are failing with `PredicateAlreadyBeingDerived`.

This is being caused by `glean derive` being called more than once for the same predicate.

This diff changes the deriver behaviour such that instead of failing, on the second invocation the deriver returns the handle of the initial derivation.

To make this work we no longer remove derivation handles from `envDerivation` when the derivation succeeds. We only remove it if the derivation fails.

This diff also removes the `PredicateAlreadyComplete` as we can now return the stats of the original successful derivation.

Reviewed By: simonmar

Differential Revision: D28193138

fbshipit-source-id: 8365e69e6499f9b4fad3193acbb023f3cba3597e