Add --schema-version to select the schema version
Summary:
Mainly for testing and migration purposes, we want to have multiple
versions of the "all" schema around. Previously the server would
always pick the latest version of "all" for resolving unversioned
types and predicates. This diff adds a new option

* `--schema-version N`

to select the version of "all" to use. The idea is that tests can use
this to select the new version during a migration.

Reviewed By: nhawkes

Differential Revision: D28289948

fbshipit-source-id: b7000c723aba0ba72d065455fec26eea3055def0