Use the glean.schema_version property on a DB when resolving unversioned predicates in a query
Summary: This is a mechnism for schema migration. The idea is that the DB associates itself with a schema version, which is used when clients make queries on the DB using unversioned predicates. This way we can add a new schema without breaking existing clients, if they're using the old DBs.

Reviewed By: ChrisKuklewicz

Differential Revision: D28324907

fbshipit-source-id: d78538a584e61327846daa73d5cda51c20e68cd8