- Fix query reordering
Summary:
A `matchType` is calculated to make sure we:
- Move point matches to the top
- Move full traversals to the bottom

`matchType` was being calculated considering as bound all variables from the initial scope **plus** all variables appearing on the left hand side of statements in the same group.

Consider the following two statement groups:
```
[ A = $123 ]          # 1
[ B = pred1 { A, _ }  # 2
, C = pred2 { B }     # 3
]
```

When calculating the `matchType` of statement 3 in the second group, given `A` and `B` are regarded as bound, the statement will be considered a point match and therefore be moved up.

This will cause a full traversal of `pred2` followed by a prefix search on `pred1`.

The original ordering was better as it consisted of a prefix search of `pred1` followed by a point match on `pred2`.

In reality, the behaviour we want is:
- If the statement is a point match using only variables from previous groups, move the statement up.
- If the statement is a point match when variables from other statements in the same group are bound, move the statement below the binding point of those variables.

Reviewed By: simonmar

Differential Revision: D28424991

fbshipit-source-id: a919e9bf2307b4ce8f07996a5035580d6d0daa2b