Back out "- Fix query reordering"
Summary:
Original commit changeset: a919e9bf2307

I am keeping the test predicates added by the original commit to decrease syncing as I will reinstitute the test that is being removed here once I propose a better reordering solution.

The perf tests show that the original changes adversely affected a category of queries that seems broader than the ones it improved.

### What is the problem?

`reorderStmtGroup` decides between two statements which should go first. However, it does not take into consideration its previous choices within the same statement group when making that decision.

With the change in D28424991 (https://github.com/facebookincubator/Glean/commit/bdfb75ee62cfe37fe4bad917e253d13bd6c8b590), given two statements, if one order meant that one of the statements would be a point-match, that order was preferred.

In the example below the algorithm looks at statements `B` and `C` (not knowing that `_131` was bound in A) and observes that if `B` comes first `C` will be a prefix match and `B` will be a full prefix traversal, however the other way around it seems like `C` would become a prefix match and `B` a point-match. It therefore opted for the second option.

This turns out to be a bad decision as the binding of `_131` earlier means `B` and `C` could be lookups, but with the order chosen `C` becomes a prefix search.

    # Before
       (1|(0|_131:hack.ClassDeclaration.4)) = D_162;              #A
        _131 = hack.ClassDeclaration.4 {_130:hack.QName.4};       #B
        _130 = hack.QName.4 {_163:hack.Name.4,_}                  #C
        )
    # After
        (1|(0|_131:hack.ClassDeclaration.4)) = D_162;             #A
        _130:hack.QName.4 = hack.QName.4 {_163:hack.Name.4,_};    #C
        _131 = hack.ClassDeclaration.4 {_130}.                    #B
        )

Reviewed By: simonmar

Differential Revision: D28507603

fbshipit-source-id: 1b9de5c910d19ea2665fef0ce8b807a0ebb814f4