- Fix editor selection in glean-shell
Summary:
The previous lookup logic was not inspecting `EDITOR` when `VISUAL` was missing, falling back to a default editor.

Considering that `ed`, although it is the Standard Text Editor, does not fulfil the mission of making it easier to write multiline queries, the new version of the code fails with an actionable error when no editor can be found.

Reviewed By: simonmar

Differential Revision: D28602657

fbshipit-source-id: 81ae7e07e008fb9e86b319a534f36b7f76098a87