deriveStored: use glean.schema_version to resolve the predicate
Summary:
We used glean.schema_version to resolve predicates in a query, but not
in the predicate passed to `deriveStored`. Yet we use unversioned
predicates with deriving a lot in tests and other places, and we need
to resolve these predicates using the DB schema to be able to do
migrations more easily.

Reviewed By: katiejots, ChrisKuklewicz

Differential Revision: D28605466

fbshipit-source-id: e4c7946b1bc7debf88cf611671a34b18239b24d7