- Simplify pollDerivation
Summary:
All relevant derivation logic from `pollDerivation` was moved to `deriveStoredImpl`, making it behave more like `deriveStored`.

With this change `updateProgress` was also renamed to `finishDerivation` to emphasise the fact that it will only be called once.

This change also fixed a race condition that could cause stored predicates with no facts to not be marked as completed.

Reviewed By: simonmar

Differential Revision: D28663733

fbshipit-source-id: 97dbb0e282c5cbfb98501a331f19ff363cc3872b