- Mark DB as broken on derivation failure
Summary:
Previously we would keep a record of a failed derivation in the environment until that error was passed on to a client that was waiting on the derivation to complete. At that point, the error was cleared and the client was free to retry deriving the predicate.

With this diff's changes we do not clear the error any longer, causing a failed derivation to prevent further attempts to derive that predicate. We also mark the database as broken, keeping a persistent record of the failure.

Reviewed By: simonmar

Differential Revision: D28715921

fbshipit-source-id: ba63030f3a4bba2fb0d9234057e2e3a362c6f5d7