Add codemarkup "variable" xrefs for Cxx
Summary:
This is barely a codemarkup query, but it does take a File and a Trace, and return the direct cxx1. spans and xref targets needed to find more xrefs in Cxx files.

The Glass client will call this as part of building all the xrefs in a file.

Ideally we would write the zip directly in Angle, then map over that to build Entities and XRefLocations. Seems useful to leave it here as specific query we can't write.

Reviewed By: simonmar

Differential Revision: D28834330

fbshipit-source-id: 2da3d6195030852324025c7b180e3fac59e2ceed