- Fix stored predicate complete dependency check
Summary:
Predicate dependencies were not being calculated correctly.
Only the return type of query statements were counted as dependencies, not predicates that were included in the statement but were not the return type.

A query line as the one below would cause the type of `A` to be considered a dependency, but not `pred1` or `pred2`:

```
(A; pred1 { A, _}) | (A; pred2 { A, _ })
```

As this check was fixed, it could be that some indexer is not compliant with the requirement that a predicate's dependency be fully derived before the predicate is itself derived.

To make indexer fixes easier, the `disable_predicate_dependency_checks` option was added to Configerator.

Reviewed By: simonmar

Differential Revision: D28642730

fbshipit-source-id: 4cd81966ff7bd8171e512973971d616012d37276