Clang indexer should keep temp files when there is an error
Summary:
The clang indexer calls buck and passes a temporary file containing the list of targets. When there is a problem sometime the buck output is sufficient for a fix. But sometimes we need the contents of the temp file to reproduce and investigate the error.

Thus when there is an error this diff will make a copy of the temp file.  These are in a subdirectory so cleanup should be easy, but I do not expect these to to take up much space.

Reviewed By: malanka

Differential Revision: D28988483

fbshipit-source-id: 86f5924de834af894b933bd40069307bf56029ad