- Add location information to variables
Summary:
The first source location annotation to be added to `SourcePat_`.

With this, some variable-related type errors now contain the location of the error.

In the future we can use this location paired with the source code to display the exact location of the error. Along the lines of the example below

```
type error in pattern

    flow.Declaration { wrong = "hello" }
                       ^^^^
    pattern: {wrong = "hello"}
    expected type: { name : flow.Name.3, loc : flow.Range.3 }
```

Reviewed By: simonmar

Differential Revision: D28830746

fbshipit-source-id: c9472374fde785ba367a003aeac11e229c0f70e5