Return ownership info from computeOwnership
Summary: We'll need to have an interface for querying the ownership information of a DB. For now, since we're not storing the ownership info in the DB itself, this is a C++ object representing the ownership info. It's mainly for wiring things together so that we can test out the ownership stuff.

Reviewed By: ChrisKuklewicz

Differential Revision: D28539144

fbshipit-source-id: 3b57bce9e1ea005d84432bc16f0151c3ad87c358