Add simple db-status tool
Summary:
It's often useful to be able to find out the status of indexing jobs
and DBs. This tool is a start; it basically does `listDatabase` on all
the write and query servers (concurrently) and displays the results,
like this:

```
$ buck run mode/opt glean/facebook/tools/db-status -- www
```

{F624551357}

There's lots more we could add to this, but I think it's already
useful to see at a glance the current status.

I plan to integrate it with the `glean` cli tool next.

Reviewed By: nhawkes

Differential Revision: D29103338

fbshipit-source-id: 2baf6a349844e577ca80baf5114c58ac16c9b74d