Add support for foreign items and unions to the Rust Glean schema.
Summary:
When testing Glean indexing across the Rust targets in `fbcode`, the Rust Glean
indexer failed due to missing support for the `ForeignFunction`,
`ForeignStatic`, and `Union` definitions that `rustc -Z save-analysis` emits.
This updates the Glean schema so that these definitions can be properly
indexed.

This diff doesn't actually index the items yet; that will be a followup.

Reviewed By: nhawkes

Differential Revision: D29150596

fbshipit-source-id: cb24ddbfcb37b55b06c65c314ff2184a39af198a