- Fix handling of failure in first branch of disjuction
Summary:
A list of statements that are obvious matches gets expanded to `[]`.

`notFalseGroups` was considering this to be a failure case, which it isn't, causing inconsistent behaviour in OR patterns with literals.
For example

```
1 = (1|2)  // succeeds
1 = (2|1)  // fails
```

This diff fixes that.

The inconsistency happened because given a disjunction where some alternatives are obvious failures and others obvious successes, the successes would be optimised away but still considered false groups within `apply (FlatDisjunction stmts)`.

This piece of code would then select the first option in the disjunction to represent all failures. If the first item was a `[]`, the query would succeed. If it was a statement with a failure sentinel it would fail.

      case filter notFalseGroups stmtss' of
        [] -> case stmtss' of
          [] -> return (FlatDisjunction [])
          (ss : _) -> return (FlatDisjunction [ss])
        some -> return (FlatDisjunction some)

Reviewed By: simonmar

Differential Revision: D29333981

fbshipit-source-id: 81e114ee7f2f6affc94d7ed4be3aa165852411db