Don't broadcast missing dbs
Summary:
Bug is:

One server has the dependencies for a db, one doesn't. A client lists dbs and finds (from the one that has the db) a complete db. It then sends the request to the server that doesn't have the db using shards.

The fix is to not broadcast dbs for which the dependencies are missing

Reviewed By: donsbot

Differential Revision: D29634023

fbshipit-source-id: 27ac9bfe38b24098f842a1004ab712bcac673e8f