Retire search.thrift:FindDecls, FindLocalGraph methods
Summary:
Part 3/N.

We can't kill search.thrift as the types are reused in glean/lib/ and are the query type of SymbolView (sigh). But we can kill the methods and remove the service from the glean server.

All traffic was originating from D29733004

Next steps would be to consolidate glean/lib/Search and symbolview? or decouple symbolview from glean/lib/*

Note: next: we need to remove the --handler flag to the glean/server
- shutdown tw/glean/server/handler = search tasks
- remove from config
- remove all reference to --handler=search

Reviewed By: malanka

Differential Revision: D29733510

fbshipit-source-id: 67804125261165ca3daa13d038aea1f39d0ab42a