Exit cleanly under buck
Summary:
If you run `buck run glean/shell` and then Ctl-C exit then buck tries to kill glean shell, but it doesn't die. The input pipe is then broken but continues to print errors to the output shell. This makes it quit if the input pipe is broken.

The catch all exceptions is fairly narrow in scope, but happy to make it more specific if you think that would be useful or print the error.

Reviewed By: simonmar

Differential Revision: D29713493

fbshipit-source-id: afbf74b7fbd2a2c9853e5196f05282a63d56092b