Fix bug treating relative bytespan arrays as absolute bytepans
Summary:
Will spotted incorrect locations for fixed xrefs in Cxx. This is a bug where Cxx "fixed" (local to file) xrefs returned by codemarkup.* were silently coerced to absolute bytespans and thus wrong locations. The first occurence of each xref would be correct but subsequent ones would have wrong offsets.

CxxFileEntityTraceFixedXRefLocations is actively affected.
CxxTargetUSes is also affected, but unused. Fixing both. This likely explains a few reports of missing or "weird" xrefs.

CxxFileEntityTraceVariableXRefLocations correctly handles this on the Haskell
side.
https://www.internalfb.com/code/fbsource/[808ea9aac4ee346d5b375ee13bff28ea61d281c5]/fbcode/glean/glass/Glean/Glass/Query/Cxx.hs?lines=141

This is arguably a type error in Angle, but is failing due to the structural coercion of RelByteSpan to ByteSpan.

Captured in regression tests.

Reviewed By: williamtwilson

Differential Revision: D30325840

fbshipit-source-id: 7bcfd989d7d6b254c9894695b5060c4413e522a5