Generate all DeclToDef xref targets for Decls in the file & trace
Summary:
For C++ compute DeclToDef for declarations in file and trace pair, filtering any duplicates.

This creates new xrefs for decls that have non-self referring Defn targets.

- xrefs generated by running code.cxx.DeclToDef on a file /trace
- these are "jump-to-definition" sorts of xrefs
- they were previously missing in glass

This exploits the new inequality primitive in Glean to filter any self-references (Decls whose Definitions are self), as we already have a decl symbol for those.

```
    Decl1 != Decl2;
```

which makes this a lot more useful as a predicate.

This will turn lots of .h decls into also xrefs to their definition, which should be quite useful.

E.g. we now find 176 new navigation targets for glean/rts/cache.h , where there were none before.
```
fbsource> :count codemarkup.CxxFileEntityTraceDeclToDefXRefLocations { file = "fbcode/glean/rts/cache.h" }

176 results, 176 facts, 31.81ms, 6442328 bytes, 44557 compiled bytes
```

This is step 1 of jump-to-defn support.

- .h decls will now be able to jump-to-defn

Step 2 is to map xrefs to decls to jump "through' the decl to the defn.
Step 3 tags to distinguish multiple xrefs to decls and defs
Step 4 invert this and do Defn to DeclFamily to jump back.

One todo: work out if expanding definition entities recursively is something we should avoid doing. Are they very expensive? We don't need the full structure to do things, just the base decl. This is _possibly_ a query that need not be recursive to be useful.

Reviewed By: nhawkes

Differential Revision: D30450331

fbshipit-source-id: 4280094fc33d4bef10f34dd7656ef04621431eee