Optimized cxx1 external xref location fetcher
Summary:
To build a code.Entity for external xref targets we do some zipping on the Glass side. We also need the span and name of the target, which is fetched pointwise by cxx1.DeclarationLocation

Instead of a map/fetch from glass, which is O(xrefs), we can do a single call and fetch all results in one, the build a lookup table from Declaration -> Location in Glass.

This should help shortern the long tail of files with large numbers of xrefs.

Reviewed By: williamtwilson

Differential Revision: D30482616

fbshipit-source-id: f188c9a22866b3dac8cec25794ff3ffc5f4e96a1