Jump from use of decl to underlying defn for "fixed" xrefs
Summary:
D30451587 adds decl to def xrefs.
This adds (ref to decl) to def xrefs. I.e. jump through the decl to the defn.

Refactor to use Location a bit earlier in the calls (these are unused methods from earlier in the stack, so safe to change type)

New xrefs are added for uses that point at the definition, rather than the decl. They can be distinguished by a) having a different target (typically a .cpp file), and the entity constructor is "defn" not "decl".

Before: xrefs to toRaw all jump to the local decl. We want the option to jump to def.

After: we get two xrefs to definitions in the file:
```
fbsource> :count codemarkup.CxxFileEntityFixedXRefLocations { file = "fbcode/skywave/skybridge/Hlc.h", xref = _ , entity = { cxx = { defn = _ }  } }

2 results, 2 facts, 18.38ms, 6463800 bytes, 37907 compiled bytes
```

Reviewed By: williamtwilson

Differential Revision: D30469477

fbshipit-source-id: 069d236ec864dad785461afd581f9a85adbb47de