- Deprecate bang for fact expansion in shell
Summary:
To avoid confusion with negation in the shell, this diff removes the use of `!` as the first query character to specify recursive fact expansion.

Recursive fact expansion is now enabled by default in the shell.

A new command was added, `:expand [on|off]`, which specifies whether facts should be recursively expanded or not.

The use of `!` to expand facts in the queries now throws an error. The idea is that after some time it would be safe to remove the error and accept `!` as the first query character meaning negation.

Reviewed By: simonmar

Differential Revision: D30585020

fbshipit-source-id: 7ad164ac946a444310982aa8725c015469a72b22