Cxx XRef facts should be indexed by FileXRefs fact
Summary:
We were keying by xmap , however, for some files there are multiple xmap facts. This was would blow things up a bit. This wasn't intentional. The intent is to pick the first "trace" of xrefs for now.

Instead, take the first cxx1.FileXRefs fact, and use that consistently. We then have the option of adding more later to gain further xrefs.

Found by profiling

Reviewed By: simonmar

Differential Revision: D30696151

fbshipit-source-id: 80e001aa9a0924b7d18a15fb8c44a119662882fc