Separate open source version of 'index' endpoint
Summary:
For now we have a stub in `github/Glean/Index.hs`.
The code is being added to `facebook/Glean/Index.hs` for the prototype.
Once the prototype works it will be decomposed into appropriate open-source and internal parts.

We will need separate OSS and internal abstractions for version control commands and for sourcing, and selecting the indexers.

Reviewed By: simonmar

Differential Revision: D30931351

fbshipit-source-id: 9a92abd729963dc745af880f4a720314ee9b8c6c