Only unwrap string
Summary:
The gshackapi treats non-struct/union predicates different to make queries look nicer. It doesn't require wrapping the predicate in a class so you can do
```
$client->query(GSHack::ClassDefinition, shape(
  'declaration' => shape('name' => shape('name' => 'ZeusGodOfThunder')),
))
```
instead of
```
$client->query(GSHack::ClassDefinition, shape(
  'declaration' => shape('name' => shape('name' => GSHack::Name('ZeusGodOfThunder'))),
))
```
This wrapping was one of the reasons the thrift api was so verbose.

Turns out this is only really common for strings, and causes complications for other types. This diff changes it so only strings get this special treatment. Its possible we may want to expand it a bit (e.g. to Option<string>) in the future

Reviewed By: simonmar

Differential Revision: D30987660

fbshipit-source-id: b1f73b5d848fcc2b96f11ec57848a5dc1b88cf4e