Add completePredicates Thrift call
Summary:
Before deriving a predicate, we need to know that the predicates on which it depends are complete. We currently have this for derived predicates, but not for non-derived predicates. In this diff I'm adding a Thrift API `completePredicates` to notify the server when all non-derived predicates are complete, so that we can start deriving.

The purpose here is twofold:
* enable correctness of deriving to be tracked and enforced,
* enable the ownership of facts in the non-derived predicates to be resolved before we start deriving. So right now I'm doing `computeOwnership` in the completion step.

There's a corresponding CLI command `glean complete`, which in the future will need to be issued before any `glean derive` commands. We can start doing that after this diff has landed and the new command is deployed.

A few other things to note here
* Nothing is currently enforced. That will come in a future diff.
* The current API is all-or-nothing, there's no support for having a mixture of complete/incomplete predicates. That means we're not going to support deriving during indexing, at least for the time being.
* In the future this API will also be used to inform the server when a manually-derived predicate is complete.

Reviewed By: lazamar

Differential Revision: D30957384

fbshipit-source-id: c154a3b9b402d508c4e5884afb166c98c8b38698