Refactor MemoryOwnership -> ComputedOwnership
Summary:
`ComputedOwnership` is the result of running `computeOwnership` - it's
a collection of owner sets and a mapping from facts to owners. This is
what will be written to the DB.

`ComputedOwnership` will also be the result of computing ownership for
a batch of derived facts, in a future diff in this stack.

Reviewed By: nhawkes

Differential Revision: D31110568

fbshipit-source-id: 0bf8725844224f8685cb5ea64e4b63ad913ccc2d