Generalise Uset to boolean expressions over sets
Summary:
The ownership of non-derived facts is represented as the disjunction of a set of units, e.g. if a fact X has owner `A || B || C`, then it would be visible if any of A, B or C is visible.

For representing the ownership of derived facts, we need to generalise the form of ownership sets, to include `&&` in addition to `||`, and to be able to nest sets so that we can represent ownerships like `(A || B) && C`.

There's a choice here: we can put our expressions into disjunctive normal form, or we can have arbitrarily nested expressions. For now I've implemented arbitrarily nested expressions because that seemed easier, but we can change this in the future if it looks like normalisation would result in fewer sets.

Reviewed By: lazamar

Differential Revision: D31139961

fbshipit-source-id: 3ee96b92d1f9d20d2fb00097601019cd5f0bd586